+++ src/components/ModalPopUp.tsx
@@ -1,10 +1,4 @@
-import {BusDetails} from "./busQueries";
-import Table from 'react-bootstrap/Table';
 import Modal from "react-bootstrap/Modal";
-import ModalBody from "react-bootstrap/ModalBody";
-import ModalHeader from "react-bootstrap/ModalHeader";
-import ModalFooter from "react-bootstrap/ModalFooter";
-import ModalTitle from "react-bootstrap/ModalTitle";
 import React, {useEffect, useState} from 'react';
 
 const ModalPopUp = ({opened, showModal, hideModal} : {opened:boolean, showModal:() => void, hideModal:() => void}): React.ReactElement => {
